+++ engine/src/mathdefs.cpp
@@ -70,14 +70,14 @@ namespace eng {
     }
 
     int VectorOrientation(const glm::ivec2& v) {
-        int orientation = int(4.f * (1.f + (std::atan2f(v.y, v.x) * glm::one_over_pi<float>())));
+        int orientation = int(4.f * (1.f + (std::atan2(v.y, v.x) * glm::one_over_pi<float>())));
         ASSERT_MSG(orientation >= 1 && orientation <= 8, "VectorOrientation - invalid conversion.");
         orientation = (8-orientation+6) % 8;
         return orientation;
     }
 
     int VectorOrientation(const glm::vec2& v) {
-        int orientation = int(4.f * (1.f + (std::atan2f(v.y, v.x) * glm::one_over_pi<float>())));
+        int orientation = int(4.f * (1.f + (std::atan2(v.y, v.x) * glm::one_over_pi<float>())));
         ASSERT_MSG(orientation >= 1 && orientation <= 8, "VectorOrientation - invalid conversion.");
         orientation = (8-orientation+6) % 8;
         return orientation;
